Restructure into multiple artifacts
Restructure project to separate client/server code and tracing/
metrics code.
