+++ .github/workflows/ci.yml
@@ -89,11 +89,11 @@ jobs:
 
       - name: Make target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: mkdir -p core/.native/target core/.js/target core/.jvm/target project/target
+        run: mkdir -p trace/client/.js/target trace/server/.js/target core/.native/target trace/internal/.js/target trace/internal/.native/target metrics/.native/target trace/server/.native/target core/.js/target trace/server/.jvm/target core/.jvm/target trace/internal/.jvm/target trace/client/.native/target metrics/.jvm/target trace/client/.jvm/target trace/core/.jvm/target trace/core/.js/target metrics/.js/target trace/core/.native/target project/target
 
       - name: Compress target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: tar cf targets.tar core/.native/target core/.js/target core/.jvm/target project/target
+        run: tar cf targets.tar trace/client/.js/target trace/server/.js/target core/.native/target trace/internal/.js/target trace/internal/.native/target metrics/.native/target trace/server/.native/target core/.js/target trace/server/.jvm/target core/.jvm/target trace/internal/.jvm/target trace/client/.native/target metrics/.jvm/target trace/client/.jvm/target trace/core/.jvm/target trace/core/.js/target metrics/.js/target trace/core/.native/target project/target
 
       - name: Upload target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
