+++ trace/client/src/main/scala/org/http4s/otel4s/middleware/trace/client/TypedClientAttributes.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2023 http4s.org
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.http4s.otel4s.middleware.trace.client
+
+import com.comcast.ip4s.Port
+import org.typelevel.otel4s.Attribute
+import org.typelevel.otel4s.semconv.attributes.ServerAttributes
+
+/** Methods for creating appropriate `Attribute`s from typed HTTP objects
+  * within an HTTP client.
+  */
+object TypedClientAttributes {
+
+  /** @return the `server.port` `Attribute` */
+  def serverPort(port: Port): Attribute[Long] =
+    ServerAttributes.ServerPort(port.value.toLong)
+}
