+++ trace/core/src/main/scala/org/http4s/otel4s/middleware/trace/ShouldTrace.scala
