+++ trace/server/src/test/scala/org/http4s/otel4s/middleware/trace/server/ServerMiddlewareTests.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.http4s.otel4s.middleware
+package org.http4s.otel4s.middleware.trace.server
 
 import cats.effect.IO
 import munit.CatsEffectSuite
